refactor(demo/tabs): Fix
